feat(proofs): update V1 service w/ DidCommMessages
