+++ packages/core/src/modules/proofs/formats/indy/IndyProofFormatService.ts
@@ -1,22 +1,49 @@
-import type { CreateProposalOptions } from '../../models/ServiceOptions'
+import type { ProposeProofFormats } from '../../models/SharedOptions'
+import type { ProofRequestOptions } from '../../protocol/v1/models'
+import type { ProofAttachmentFormat } from '../models/ProofAttachmentFormat'
 import type {
   CreatePresentationOptions,
+  CreateProposalOptions,
+  CreateRequestAttachmentOptions,
   CreateRequestOptions,
   ProcessProposalOptions,
   ProcessRequestOptions,
-  ProofAttachmentFormat,
-  ProofFormatSpec,
-} from '../ProofFormatService'
+} from '../models/ProofFormatServiceOptions'
+import type { ProofFormatSpec } from '../models/ProofFormatSpec'
 
 import { Attachment, AttachmentData } from '../../../../decorators/attachment/Attachment'
+import { AriesFrameworkError } from '../../../../error/AriesFrameworkError'
 import { JsonEncoder } from '../../../../utils/JsonEncoder'
+import { ProofRequest } from '../../protocol/v1/models'
 import { ProofFormatService } from '../ProofFormatService'
 import { ATTACHMENT_FORMAT } from '../ProofFormats'
 
 export class IndyProofFormatService extends ProofFormatService {
+  private createRequestAttachment(options: CreateRequestAttachmentOptions) {
+    const format: ProofFormatSpec = this.getFormatIdentifier(options.messageType)
+
+    const request = new ProofRequest(options.proofRequestOptions)
+
+    const attachment = new Attachment({
+      id: options.attachId,
+      mimeType: 'application/json',
+      data: new AttachmentData({
+        base64: JsonEncoder.toBase64(request),
+      }),
+    })
+    return { format, attachment }
+  }
+
   public createProposal(options: CreateProposalOptions): ProofAttachmentFormat {
-    // Handle format in service
-    throw new Error('Method not implemented.')
+    if (!options.formats.indy) {
+      throw Error('Indy format missing')
+    }
+
+    return this.createRequestAttachment({
+      attachId: options.attachId,
+      messageType: options.messageType,
+      proofRequestOptions: options.formats.indy,
+    })
   }
 
   public processProposal(options: ProcessProposalOptions): void {
@@ -24,17 +51,17 @@ export class IndyProofFormatService extends ProofFormatService {
   }
 
   public createRequest(options: CreateRequestOptions): ProofAttachmentFormat {
-    const format: ProofFormatSpec = this.getFormatIdentifier(options.messageType)
+    if (!options.formats.indy) {
+      throw new AriesFrameworkError(
+        'Unable to get requested credentials for proof request. No proof request message was found or the proof request message does not contain an indy proof request.'
+      )
+    }
 
-    const { attachId, proofRequest } = options
-    const attachment = new Attachment({
-      id: attachId ? attachId : undefined,
-      mimeType: 'application/json',
-      data: new AttachmentData({
-        base64: JsonEncoder.toBase64(proofRequest),
-      }),
+    return this.createRequestAttachment({
+      attachId: options.attachId,
+      messageType: options.messageType,
+      proofRequestOptions: options.formats.indy,
     })
-    return { format, attachment }
   }
 
   public processRequest(options: ProcessRequestOptions): void {
