+++ packages/core/src/modules/proofs/formats/models/ProofAttachmentFormat.ts
@@ -1,0 +1,7 @@
+import type { Attachment } from '../../../../decorators/attachment/Attachment'
+import type { ProofFormatSpec } from './ProofFormatSpec'
+
+export interface ProofAttachmentFormat {
+  format: ProofFormatSpec
+  attachment: Attachment
+}
