--- packages/core/src/modules/proofs/protocol/v2/formats/MessageTypes.ts
@@ -1,11 +1,0 @@
-export const SPEC_URI =
-  'https://github.com/hyperledger/aries-rfcs/tree/eace815c3e8598d4a8dd7881d8c731fdb2bcc0aa/features/0454-present-proof-v2'
-
-// Message types
-export const PRES_20_PROPOSAL = 'present-proof/2.0/propose-presentation'
-export const PRES_20_REQUEST = 'present-proof/2.0/request-presentation'
-export const PRES_20 = 'present-proof/2.0/presentation'
-export const PRES_20_ACK = 'present-proof/2.0/ack'
-export const PRES_20_PROBLEM_REPORT = 'present-proof/2.0/problem-report'
-
-export const PROTOCOL_PACKAGE = 'aries_cloudagent.protocols.present_proof.v2_0'
