+++ packages/core/src/modules/proofs/repository/ProofRecord.ts
@@ -1,32 +1,23 @@
 import type { TagsBase } from '../../../storage/BaseRecord'
 import type { AutoAcceptProof } from '../models/ProofAutoAcceptType'
+import type { ProofProtocolVersion } from '../models/ProofProtocolVersion'
 import type { ProofState } from '../models/ProofState'
-import type { V2ProposalPresentationMessage } from '../v2/messages/V2ProposalPresentationMessage'
-
-import { Type } from 'class-transformer'
 
 import { AriesFrameworkError } from '../../../error'
 import { BaseRecord } from '../../../storage/BaseRecord'
 import { uuid } from '../../../utils/uuid'
-import { ProposePresentationMessage, RequestPresentationMessage, PresentationMessage } from '../protocol/v1/messages'
 
 export interface ProofRecordProps {
   id?: string
   createdAt?: Date
-
+  protocolVersion: ProofProtocolVersion
   isVerified?: boolean
   state: ProofState
   connectionId?: string
   threadId: string
-  presentationId?: string
   tags?: CustomProofTags
   autoAcceptProof?: AutoAcceptProof
   errorMessage?: string
-
-  // message data
-  proposalMessage?: ProposePresentationMessage | V2ProposalPresentationMessage
-  requestMessage?: RequestPresentationMessage
-  presentationMessage?: PresentationMessage
 }
 
 export type CustomProofTags = TagsBase
@@ -39,19 +30,11 @@ export type DefaultProofTags = {
 export class ProofRecord extends BaseRecord<DefaultProofTags, CustomProofTags> {
   public connectionId?: string
   public threadId!: string
+  public protocolVersion!: ProofProtocolVersion
   public isVerified?: boolean
-  public presentationId?: string
   public state!: ProofState
   public autoAcceptProof?: AutoAcceptProof
   public errorMessage?: string
-
-  // message data -- this goes in a separate DIDComm message record
-  @Type(() => ProposePresentationMessage)
-  public proposalMessage?: ProposePresentationMessage | V2ProposalPresentationMessage
-  @Type(() => RequestPresentationMessage)
-  public requestMessage?: RequestPresentationMessage
-  @Type(() => PresentationMessage)
-  public presentationMessage?: PresentationMessage
 
   public static readonly type = 'ProofRecord'
   public readonly type = ProofRecord.type
@@ -62,14 +45,13 @@ export class ProofRecord extends BaseRecord<DefaultProofTags, CustomProofTags> {
     if (props) {
       this.id = props.id ?? uuid()
       this.createdAt = props.createdAt ?? new Date()
-      this.proposalMessage = props.proposalMessage
-      this.requestMessage = props.requestMessage
-      this.presentationMessage = props.presentationMessage
+      this.protocolVersion = props.protocolVersion
+
       this.isVerified = props.isVerified
       this.state = props.state
       this.connectionId = props.connectionId
       this.threadId = props.threadId
-      this.presentationId = props.presentationId
+
       this.autoAcceptProof = props.autoAcceptProof
       this._tags = props.tags ?? {}
       this.errorMessage = props.errorMessage
